% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa_base.R
\name{calculate}
\alias{calculate}
\title{Calculate the kappa index proposed in the Sun W, Wang J, Fang Y. Consistent selection of tuning parameters via variable selection stability. The Journal of Machine Learning Research, 2013, 14(1): 3419-3440.}
\usage{
calculate(result1, result2)
}
\arguments{
\item{result1}{a vector contains the variable selected from the sub dataset 1}

\item{result2}{a vector contains the variable selected from the sub dataset 2}
}
\value{
The calculate Kappa index from result1 and result2
}
\description{
Calculate the kappa index proposed in the Sun W, Wang J, Fang Y. Consistent selection of tuning parameters via variable selection stability. The Journal of Machine Learning Research, 2013, 14(1): 3419-3440.
}
