% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVXR_base.R
\name{CVXR_Quantile}
\alias{CVXR_Quantile}
\title{The Quantile loss function with RKHS ridge penalty}
\usage{
CVXR_Quantile(y, K, lambda, tau = 0.5, print_loss = FALSE)
}
\arguments{
\item{y}{is the response variable.}

\item{K}{is the Kernel Matrix generated from the package kernlab. K@.Data is a matrix contains the data.}

\item{lambda}{is the tunable parameter of RKHS ridge penalty.}

\item{tau}{is the quantile level we choose. tau = 0.5 by default.}

\item{print_loss}{a logic value to decide whether print loss or not}
}
\description{
The Quantile loss function with RKHS ridge penalty
}
