% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVXR_base.R
\name{CVXR_SVC}
\alias{CVXR_SVC}
\title{The SVM-Classification loss function with RKHS ridge penalty}
\usage{
CVXR_SVC(y, K, lambda, print_loss = FALSE)
}
\arguments{
\item{y}{is the response variable.}

\item{K}{is the Kernel Matrix generated from the package kernlab. K@.Data is a matrix contains the data.}

\item{lambda}{is the tunable parameter of RKHS ridge penalty.}

\item{print_loss}{a logic value to decide whether print loss or not}
}
\description{
The SVM-Classification loss function with RKHS ridge penalty
}
