% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL.R
\name{DSL_base_screening}
\alias{DSL_base_screening}
\title{The function calculate the gradient of RKHS regression for L2, Quantile, Huber, Hinge and Logistic loss. The combination of the DSL_calculate_alpha and DSL_calculate_gradient taht is a Cpp function in the Calculate_GradienKeh.cpp}
\usage{
DSL_base_screening(
  X,
  Y,
  Loss = "L2",
  lambda = 1,
  Tau = 0.5,
  Kernel = "Gaussian",
  isKernelScale = FALSE,
  print_loss = TRUE,
  only_gradient = TRUE
)
}
\arguments{
\item{X}{is the covariates matrix.}

\item{Y}{is the response vector.}

\item{Loss}{the loss choosed among L2 Check Huber Logistic SVM-Classification}

\item{lambda}{is the tunable parameter of RKHS ridge penalty.}

\item{Tau}{is the quantile level we choose. tau = 0.5 by default. for other loss, tau is useless.}

\item{Kernel}{only Gaussian Kernel supported now.}

\item{isKernelScale}{parameter for generating the kernel matrix deployed by the kernlab package. default FALSE and the performance is satisfactory.}

\item{print_loss}{a logic value to decide whether print loss or not}

\item{only_gradient}{only a list contains gradient returned or a list contains more information returned.}
}
\description{
The function calculate the gradient of RKHS regression for L2, Quantile, Huber, Hinge and Logistic loss. The combination of the DSL_calculate_alpha and DSL_calculate_gradient taht is a Cpp function in the Calculate_GradienKeh.cpp
}
