% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL.R
\name{DSL_base_weak_heredity}
\alias{DSL_base_weak_heredity}
\title{Under weak heredity setting, the function calculate the second order gradient of RKHS regression for L2, Quantile, Huber, Hinge and Logistic loss}
\usage{
DSL_base_weak_heredity(
  X,
  Y,
  Loss = "L2",
  lambda = 1,
  Tau = 0.5,
  Ind = NA,
  Kernel = "Gaussian",
  isKernelScale = FALSE,
  print_loss = TRUE
)
}
\arguments{
\item{X}{is the covariates matrix.}

\item{Y}{is the response vector.}

\item{Loss}{loss function defined: L2, Check,Huber,SVM_Classification, Logistic}

\item{lambda}{is the tunable parameter of RKHS ridge penalty.}

\item{Tau}{is the quantile level we choose. tau = 0.5 by default. for other loss, tau is useless.}

\item{Ind}{the vector of the index contains the main effect selected.}

\item{Kernel}{only Gaussian Kernel supported now.}

\item{isKernelScale}{parameter for generating the kernel matrix deployed by the kernlab package. default FALSE and the performance is satisfactory.}

\item{print_loss}{a logic value to decide whether print loss or not}
}
\description{
Under weak heredity setting, the function calculate the second order gradient of RKHS regression for L2, Quantile, Huber, Hinge and Logistic loss
}
