% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_order2_vec}
\alias{get_order2_vec}
\title{The function postpropose the interaction selection problem}
\usage{
get_order2_vec(bool_mat_with_name)
}
\arguments{
\item{bool_mat_with_name}{translate the bool matrix to selected interaction term}
}
\value{
return a vector  contains the interaction selected.
}
\description{
The function postpropose the interaction selection problem
}
