% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSLM-package.R
\docType{package}
\name{GSLM-package}
\alias{GSLM-package}
\alias{_PACKAGE}
\title{GSLM}
\description{
A package for Structure learning via unstructured kernel-based M-regression in Reproducing Kernel Hilbert Space
\tabular{ll}{
  Package: \tab GSLM\cr
  Type: \tab Package\cr
  Version: \tab 1.0.0\cr
  Date: \tab 2021-04-28\cr
  License: \tab GPL-2\cr
  LazyLoad: \tab yes\cr
}
}
\references{
The meanuscripts
}
\author{
Xin He, Yeheng Ge and Xingdong Feng \cr
Maintainers: Yeheng Ge <geyh96@163.sufe.edu.cn>;
}
