% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa_base.R
\name{Kapp.cal}
\alias{Kapp.cal}
\title{calculate the kappa index proposed in the Sun(2013). the kap.lam with largest return value are selected.}
\usage{
Kapp.cal(kap.re1, kap.re2, kap.lam)
}
\arguments{
\item{kap.re1}{a matrix of b by p . b the the time of simulation and p is the dimension of covariates . kap.re1[i,] contains the variable selected from the sub dataset 1 in the i'th loop}

\item{kap.re2}{a matrix of b by p . b the the time of simulation and p is the dimension of covariates . kap.re1[i,] contains the variable selected from the sub dataset 2 in the i'th loop}

\item{kap.lam}{the candidate of threshold used in the GSLM framework to get exact variable selection.}
}
\value{
The calculate Kappa index corresponding to the kap.lam
}
\description{
calculate the kappa index proposed in the Sun(2013). the kap.lam with largest return value are selected.
}
