% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa_fun.R
\name{Kappa_DSL}
\alias{Kappa_DSL}
\title{The function select the RKHS ridge penalty parameter and threshold via stability selection proposed by SUN(2013)}
\usage{
Kappa_DSL(
  X,
  Y,
  Loss = "L2",
  kappa_Tau = 0.5,
  lambdas = seq(0, 2, 0.5),
  Kernel = "Gaussian",
  kappa_time = 5,
  kappa_thres = 10^(seq(-1.7, -1.2, 0.1)),
  d_kapparatio = 1,
  isKernelScale = FALSE
)
}
\arguments{
\item{X}{is the covariates matrix.}

\item{Y}{is the response vector.}

\item{Loss}{loss function defined: L2, Check,Huber,SVM_Classification, Logistic}

\item{kappa_Tau}{is the quantile level we choose. tau = 0.5 by default. for other loss, tau is useless.}

\item{lambdas}{the vector  contains the candiate of the RKHS ridge parameter lambda.}

\item{Kernel}{only Gaussian Kernel supported now.}

\item{kappa_time}{the times of the data split and kappa calculated}

\item{kappa_thres}{the candidates of the thresholding value}

\item{d_kapparatio}{the alpha_n in the paper SUN (2013) algorithm Step4}

\item{isKernelScale}{parameter for generating the kernel matrix deployed by the kernlab package. default FALSE and the performance is satisfactory.}
}
\value{
return a list contains the lambda and threshold selected.
}
\description{
The function select the RKHS ridge penalty parameter and threshold via stability selection proposed by SUN(2013)
}
