% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_with_order}
\alias{paste_with_order}
\title{The function postpropose the interaction selection problem}
\usage{
paste_with_order(str1, str2)
}
\arguments{
\item{str1}{a str like "X1"}

\item{str2}{a str like "X2"}
}
\value{
return a vector  like "X1X2"
}
\description{
The function postpropose the interaction selection problem
}
