% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_to_var}
\alias{str_to_var}
\title{The function postpropose the interaction selection problem}
\usage{
str_to_var(str)
}
\arguments{
\item{str}{a str like "X1X2"}
}
\value{
return a vector  contains the interaction selected.
}
\description{
The function postpropose the interaction selection problem
}
